+++ apps/root/src/pages/dca/positions/components/positions-list/current-positions/components/create-position-box/index.tsx
